fix(reporter): utc date handling and update dependencies (#562)
Fix: Always assume UTC time in NswgReporter

NswgReporter would use whatever time zone of the computer on which it is
running. This meant that tests passed for people in time zones with no
offset from UTC or a positive offset, but tests failed for people (like
me) in time zones where the offset is negative. (Dates that were
expected to be 1970-01-01 were instead 1969-12-31.)

Use UTC-specific functions so that dates are handled consistently.
Before this, dates could be shifted back one day on people in time zones
with negative offsets.